fix some issues reported by golangci-lint

+++ zlog/logfmt_encoder.go
@@ -177,7 +177,7 @@ func (enc *logfmtEncoder) AppendByteString(value []byte) {
 
 func (enc *logfmtEncoder) AppendComplex128(value complex128) {
 	// Cast to a platform-independent, fixed-size type.
-	r, i := float64(real(value)), float64(imag(value))
+	r, i := real(value), imag(value)
 	enc.buf.AppendFloat(r, 64)
 	enc.buf.AppendByte('+')
 	enc.buf.AppendFloat(i, 64)
@@ -472,7 +472,7 @@ func (enc *literalEncoder) AppendByteString(value []byte) {
 func (enc *literalEncoder) AppendComplex128(value complex128) {
 	enc.addSeparator()
 	// Cast to a platform-independent, fixed-size type.
-	r, i := float64(real(value)), float64(imag(value))
+	r, i := real(value), imag(value)
 	enc.buf.AppendFloat(r, 64)
 	enc.buf.AppendByte('+')
 	enc.buf.AppendFloat(i, 64)
